requirements.txt
streamlit
requests
pandas
plotly
numpy
python-dotenv